+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- List of items is now retrieved from `OrderManager`.
+- `updateItem` optimistically updates Subtotal and Total values in order form.
+
+### Removed
+
+- `itemList` from the API.
+
 ## [0.1.0] - 2019-08-23
 
 ### Added